+++ npm-client/types/index.ts
@@ -3,7 +3,6 @@ export * from "./default_price_ladder";
 export * from "./errors";
 export * from "./order";
 export * from "./market";
-export * from "./market_matching_queue";
 export * from "./get_account";
 export * from "./market_position";
 export * from "./matching_pool";
